FileLogger improvement
Formerly the save function aimed just to close the files to save the contents. Now the content is written little by little in a character string, once the save method has been executed, this content is written and the file is closed to save it, the content is therefore reset to 0 at each save.